Pull request comments
Functions moved to helpers file.
